% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeak.R
\name{callPeak}
\alias{callPeak}
\title{Call peaks}
\usage{
callPeak(bamfile, index = bamfile, inputBam = NULL, inputIndex = inputBam,
  txdb, genome, upstream = 50000, downstream = 50000, ideaPeakWidth = 200,
  FDRfilter = 0.05, direction = c("over", "less"))
}
\arguments{
\item{bamfile}{bam file name}

\item{index}{index file name}

\item{inputBam}{input bam file name}

\item{inputIndex}{index file name for input}

\item{txdb}{TxDb object}

\item{genome}{BSgenome object}

\item{upstream}{peak scan start position from upstream of gene}

\item{downstream}{peak scan end position till downstream of gene}

\item{ideaPeakWidth}{ideally peak width, eg, ATAC-seq: 200bp}

\item{FDRfilter}{cut off value of FDR.}

\item{direction}{over or less. In most case, default over is OK. 
In some cases, such as cohesion, less may be what you want to try.}
}
\value{
a GRanage object indicates the peaks.
}
\description{
Calling peaks from bam files
}
\examples{
path <- system.file("extdata", package = "diffPeaks", mustWork = TRUE)
bamfiles <- dir(path, "bam$", full.names = TRUE)
bamfile <- bamfiles[1]
index <- bamfile
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
library(BSgenome.Hsapiens.UCSC.hg19)
genome <- Hsapiens
upstream <- downstream <- 50000L
ideaPeakWidth <- 200
direction <- "over"
x <- callPeak(bamfile, txdb=txdb, genome=genome)
}
