% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMHpeaks.R
\name{CMHpeaks}
\alias{CMHpeaks}
\title{differential binding analysis by Cochran-Mantel-Haenszel Test}
\usage{
CMHpeaks(counts, conditionA, conditionB, ...)
}
\arguments{
\item{counts}{output of \link{countTable}}

\item{conditionA, conditionB}{condition A and B. Comparison will be A-B.}

\item{...}{not used.}
}
\value{
an object of \link[GenomicRanges]{GRanges}
}
\description{
use Cochran-Mantel-Haenszel Test to test differential peaks
}
\examples{
path <- system.file("extdata", package = "diffPeaks", mustWork = TRUE)
bamfiles <- dir(path, "bam$")
peaks <- dir(path, "bed$")
p <- mergePeaks(file.path(path, peaks))
colData <- DataFrame(samples=bamfiles, condition=sub(".rep..bam", "", bamfiles), 
                     pairs=rep(c(1,2,3), 2))
cnt <- countTable(p, file.path(path, bamfiles), colData)
CMHpeaks(cnt, conditionA="inj", conditionB="uni")
}
\author{
Jianhong Ou
}
