% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeak.R
\name{count5end}
\alias{count5end}
\title{count reads by regions}
\usage{
count5end(bamfile, index, region, ideaPeakWidth)
}
\arguments{
\item{bamfile}{bam file name}

\item{index}{index file name}

\item{region}{an object of GRanges for counting.}

\item{ideaPeakWidth}{ideally peak width, eg, ATAC-seq: 200bp}
}
\value{
an object of GRanges with counting
}
\description{
help function for callPeak
}
