% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTable.R
\name{countTable}
\alias{countTable}
\title{Counting reads in features}
\usage{
countTable(features, bamfiles, colData, samples = bamfiles,
  windowSize = 200L, ...)
}
\arguments{
\item{features}{an object of \link[GenomicRanges]{GRanges}}

\item{bamfiles}{filenames of aligned reads}

\item{colData}{a DataFrame or data.frame with at least a single column. 
Rows of colData correspond to bamfiles}

\item{samples}{names of samples, could be same length with bamfiles}

\item{windowSize}{the size of windows for counts}

\item{...}{parameters could be passed to \link[GenomicAlignments]{summarizeOverlaps}}
}
\value{
a list of \link[SummarizedExperiment]{RangedSummarizedExperiment}
}
\description{
Counting reads in giving features.
}
\examples{
path <- system.file("extdata", package = "diffPeaks", mustWork = TRUE)
bamfiles <- dir(path, "bam$")
peaks <- dir(path, "bed$")
p <- mergePeaks(file.path(path, peaks))
colData <- DataFrame(samples=bamfiles, condition=sub(".rep..bam", "", bamfiles))
cnt <- countTable(p, file.path(path, bamfiles), colData)
}
\author{
Jianhong Ou
}
