% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBFpeaks.R
\name{DBFpeaks}
\alias{DBFpeaks}
\title{differential binding analysis}
\usage{
DBFpeaks(counts, ...)
}
\arguments{
\item{counts}{output of \link{countTable}}

\item{...}{parameters could be passed to \link[DESeq2]{DESeqDataSet}}
}
\value{
an list of \link[GenomicRanges]{GRanges}
}
\description{
use DBF to test differential peaks
}
\examples{
path <- system.file("extdata", package = "diffPeaks", mustWork = TRUE)
bamfiles <- dir(path, "bam$")
peaks <- dir(path, "bed$")
p <- mergePeaks(file.path(path, peaks))
colData <- DataFrame(samples=bamfiles, condition=sub(".rep..bam", "", bamfiles))
cnt <- countTable(p, file.path(path, bamfiles), colData)
DBFpeaks(cnt, design= ~condition)

}
\author{
Jianhong Ou
}
