% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEpeaks.R
\name{DEpeaks}
\alias{DEpeaks}
\title{differential binding analysis}
\usage{
DEpeaks(counts, ...)
}
\arguments{
\item{counts}{output of \link{countTable}}

\item{...}{parameters could be passed to \link[DESeq2]{DESeqDataSet}}
}
\value{
an object of \link[GenomicRanges]{GRanges}
}
\description{
use DEseq2 to test differential peaks
}
\examples{
path <- system.file("extdata", package = "diffPeaks", mustWork = TRUE)
bamfiles <- dir(path, "bam$")
peaks <- dir(path, "bed$")
p <- mergePeaks(file.path(path, peaks))
colData <- DataFrame(samples=bamfiles, condition=sub(".rep..bam", "", bamfiles))
cnt <- countTable(p, file.path(path, bamfiles), colData)
DEpeaks(cnt, design= ~condition)
}
\author{
Jianhong Ou
}
