\name{DiffPeaks-package}
\alias{DiffPeaks-package}
\alias{DiffPeaks}
\docType{package}
\title{Compute the difference between two peak tracks by alignment}
\description{Chromatin immunoprecipitation DNA sequencing results in genomic
tracks that show enriched regions or peaks where proteins are bound.
This package implements fast C++ code that aligns two different peak tracks
and outputs the differences.}
\details{
\tabular{ll}{Package: \tab DiffPeaks\cr
Maintainer: \tab Toby Dylan Hocking <tdhock5@gmail.com>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 2014.5.27\cr
License: \tab GPL-3\cr
Title: \tab Compute the difference between two peak tracks by alignment\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


