% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTpeaks.R
\name{FTpeaks}
\alias{FTpeaks}
\title{differential binding analysis by Fisher's exact test}
\usage{
FTpeaks(counts, conditionA, conditionB, ...)
}
\arguments{
\item{counts}{output of \link{countTable}}

\item{conditionA, conditionB}{condition A and B. Comparison will be A-B.}

\item{...}{not used.}
}
\value{
an object of \link[GenomicRanges]{GRanges}
}
\description{
use Fishter's exact test to test differential peaks
}
\examples{
path <- system.file("extdata", package = "diffPeaks", mustWork = TRUE)
bamfiles <- dir(path, "bam$")
peaks <- dir(path, "bed$")
p <- mergePeaks(file.path(path, peaks))
colData <- DataFrame(samples=bamfiles, condition=sub(".rep..bam", "", bamfiles))
cnt <- countTable(p, file.path(path, bamfiles), colData)
FTpeaks(cnt, conditionA="inj", conditionB="uni")
}
\author{
Jianhong Ou
}
