% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTable.R
\name{isCountTable}
\alias{isCountTable}
\title{is output of countTable}
\usage{
isCountTable(counts, error = "counts must be output of countTable!")
}
\arguments{
\item{counts}{the object to be checked}

\item{error}{the error message}
}
\description{
check an object is output of \link{countTable} or not.
}
