% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePeaks.R
\name{mergePeaks}
\alias{mergePeaks}
\title{Merge BED files}
\usage{
mergePeaks(peaks, maxPeakWidth = 5000, ...)
}
\arguments{
\item{peaks}{BED filenames which indicate the peaks or an GRangesList 
or GRanges object could be used as input peaks.}

\item{maxPeakWidth}{maximal peak width. If greater than maxPeakWidth,
the peak will be divided in half to fit the parameter.}

\item{...}{parameters could be passed to \link[rtracklayer]{import}}
}
\value{
an object of \link[GenomicRanges]{GRanges}
}
\description{
Merge multiple peak files
}
\examples{
path <- system.file("extdata", package = "diffPeaks", mustWork = TRUE)
peaks <- dir(path, "bed$")
p <- mergePeaks(file.path(path, peaks))
}
\author{
Jianhong Ou
}
