% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPeaks.R
\docType{package}
\name{diffPeaks-package}
\alias{diffPeaks}
\alias{diffPeaks-package}
\title{Quantifying Differential Features}
\description{
Find the difference in the coverage of given genomic ranges. 
The difference including quantity difference and shape difference in peaks.
}
